+++ src/components/AdjustButtonContainer/index.tsx
@@ -1,13 +1,76 @@
 import * as S from './style';
 
-import { ReactNode } from 'react';
+import { useContext, useEffect, useState } from 'react';
+
+import { UseCartItemsContext } from '../../App';
+import AdjustButton from '../AdjustButton';
+import QuantityText from '../QuantityText';
+import Spinner from '../Spinner';
+
+import { CartItem } from '../../types/cart';
+import { MINUS_BUTTON, PLUS_BUTTON, REMOVE_BUTTON } from '../../assets/images';
 
 interface AdjustButtonContainerProps {
-  children: ReactNode;
+  cartItem: CartItem;
 }
 
-const AdjustButtonContainer = ({ children }: AdjustButtonContainerProps) => {
-  return <S.AdjustButtonContainer>{children}</S.AdjustButtonContainer>;
+const AdjustButtonContainer = ({ cartItem }: AdjustButtonContainerProps) => {
+  const { deleteCartItem, adjustCartItemQuantity } = useContext(UseCartItemsContext);
+  const [isLoading, setIsLoading] = useState(false);
+
+  const handleDecreaseCartItemQuantity = () => {
+    setIsLoading(true);
+
+    if (cartItem?.quantity === 1) {
+      return deleteCartItem.mutate(cartItem.id, {
+        onError: () => setIsLoading(false),
+      });
+    }
+
+    if (cartItem)
+      adjustCartItemQuantity.mutate(
+        { cartItemId: cartItem.id, quantity: cartItem.quantity - 1 },
+        {
+          onError: () => setIsLoading(false),
+        },
+      );
+  };
+
+  const handleIncreaseCartItemQuantity = () => {
+    setIsLoading(true);
+
+    if (cartItem)
+      adjustCartItemQuantity.mutate(
+        { cartItemId: cartItem.id, quantity: cartItem.quantity + 1 },
+        {
+          onError: () => setIsLoading(false),
+        },
+      );
+  };
+
+  useEffect(() => {
+    setIsLoading(false);
+  }, [cartItem.quantity]);
+
+  return (
+    <S.AdjustButtonContainer>
+      <AdjustButton onClick={handleDecreaseCartItemQuantity}>
+        <img
+          src={cartItem.quantity === 1 ? REMOVE_BUTTON : MINUS_BUTTON}
+          alt={cartItem.quantity === 1 ? '삭제' : '빼기'}
+        />
+      </AdjustButton>
+      {!isLoading && <QuantityText>{cartItem.quantity}</QuantityText>}
+      {isLoading && (
+        <QuantityText>
+          <Spinner />
+        </QuantityText>
+      )}
+      <AdjustButton onClick={handleIncreaseCartItemQuantity}>
+        <img src={PLUS_BUTTON} alt={'더하기'} />
+      </AdjustButton>
+    </S.AdjustButtonContainer>
+  );
 };
 
 export default AdjustButtonContainer;
