+++ src/components/ProductItem/index.tsx
@@ -4,11 +4,9 @@ import { useContext, useEffect, useState } from 'react';
 
 import { UseCartItemsContext } from '../../App';
 import AdjustButtonContainer from '../AdjustButtonContainer';
-import AdjustButton from '../AdjustButton';
 import Spinner from '../Spinner';
-import QuantityText from '../QuantityText';
 
-import { ADD_TO_CART, MINUS_BUTTON, PLUS_BUTTON, REMOVE_BUTTON } from '../../assets/images';
+import { ADD_TO_CART } from '../../assets/images';
 
 interface ProductItemProps {
   id: number;
@@ -18,8 +16,7 @@ interface ProductItemProps {
 }
 
 const ProductItem = ({ id, imageUrl, name, price }: ProductItemProps) => {
-  const { getCartItems, addCartItem, deleteCartItem, adjustCartItemQuantity } =
-    useContext(UseCartItemsContext);
+  const { getCartItems, addCartItem } = useContext(UseCartItemsContext);
   const [isLoading, setIsLoading] = useState(false);
 
   const cartItem = getCartItems.data
@@ -35,36 +32,6 @@ const ProductItem = ({ id, imageUrl, name, price }: ProductItemProps) => {
     addCartItem.mutate(id, {
       onError: () => setIsLoading(false),
     });
-  };
-
-  const handleDecreaseCartItemQuantity = () => {
-    setIsLoading(true);
-
-    if (cartItem?.quantity === 1) {
-      return deleteCartItem.mutate(cartItem.id, {
-        onError: () => setIsLoading(false),
-      });
-    }
-
-    if (cartItem)
-      adjustCartItemQuantity.mutate(
-        { cartItemId: cartItem.id, quantity: cartItem.quantity - 1 },
-        {
-          onError: () => setIsLoading(false),
-        },
-      );
-  };
-
-  const handleIncreaseCartItemQuantity = () => {
-    setIsLoading(true);
-
-    if (cartItem)
-      adjustCartItemQuantity.mutate(
-        { cartItemId: cartItem.id, quantity: cartItem.quantity + 1 },
-        {
-          onError: () => setIsLoading(false),
-        },
-      );
   };
 
   useEffect(() => {
@@ -91,25 +58,7 @@ const ProductItem = ({ id, imageUrl, name, price }: ProductItemProps) => {
               {isLoading && <Spinner />}
             </S.AddButton>
           )}
-          {isInCart && (
-            <AdjustButtonContainer>
-              <AdjustButton onClick={handleDecreaseCartItemQuantity}>
-                <img
-                  src={cartItem.quantity === 1 ? REMOVE_BUTTON : MINUS_BUTTON}
-                  alt={cartItem.quantity === 1 ? '삭제' : '빼기'}
-                />
-              </AdjustButton>
-              {!isLoading && <QuantityText>{cartItem.quantity}</QuantityText>}
-              {isLoading && (
-                <QuantityText>
-                  <Spinner />
-                </QuantityText>
-              )}
-              <AdjustButton onClick={handleIncreaseCartItemQuantity}>
-                <img src={PLUS_BUTTON} alt={'더하기'} />
-              </AdjustButton>
-            </AdjustButtonContainer>
-          )}
+          {isInCart && <AdjustButtonContainer cartItem={cartItem} />}
         </S.ButtonContainer>
       </S.InformationContainer>
     </S.ProductItem>
